JBPM-5601 - Move KieServicesHttpException and KieServicesException to kie-server-api module
Moving exceptions to API module so that it is exported and accessible to GWT/Errai
