+++ kie-server-parent/kie-server-api/src/main/resources/ErraiApp.properties
@@ -33,4 +33,6 @@
 errai.marshalling.serializableTypes=org.kie.server.api.model.Message \
 org.kie.server.api.model.ReleaseId \
 org.kie.server.api.model.Severity \
-org.kie.server.api.model.KieContainerStatus
+org.kie.server.api.model.KieContainerStatus \
+org.kie.server.api.exception.KieServicesException \
+org.kie.server.api.exception.KieServicesHttpException
