+++ kie-server-parent/kie-server-api/src/main/resources/org/kie/server/api/KieServerAPI.gwt.xml
@@ -19,4 +19,9 @@
     <include name="KieServerInfo.java"/>
   </source>
 
+  <source path="exception">
+    <include name="KieServicesException.java"/>
+    <include name="KieServicesHttpException.java"/>
+  </source>
+
 </module>
