+++ kie-server-parent/kie-server-remote/kie-server-client/src/test/java/org/kie/server/client/KieServicesClientErrorHandlingTest.java
@@ -20,6 +20,8 @@ import org.hamcrest.Description;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
+import org.kie.server.api.exception.KieServicesException;
+import org.kie.server.api.exception.KieServicesHttpException;
 import org.kie.server.api.marshalling.MarshallingException;
 import org.kie.server.api.marshalling.MarshallingFormat;
 import org.kie.server.client.impl.KieServicesClientImpl;
