+++ kie-server-parent/kie-server-tests/kie-server-integ-tests-all/src/test/java/org/kie/server/integrationtests/jbpm/ImageServiceIntegrationTest.java
@@ -18,7 +18,7 @@ package org.kie.server.integrationtests.jbpm;
 import org.junit.BeforeClass;
 import org.junit.Test;
 import org.kie.server.api.model.ReleaseId;
-import org.kie.server.client.KieServicesException;
+import org.kie.server.api.exception.KieServicesException;
 
 import static org.junit.Assert.*;
 import org.kie.server.integrationtests.shared.KieServerDeployer;
