+++ kie-server-parent/kie-server-tests/kie-server-integ-tests-all/src/test/java/org/kie/server/integrationtests/jbpm/PerProcessInstanceWorkItemIntegrationTest.java
@@ -25,7 +25,7 @@ import org.junit.BeforeClass;
 import org.junit.Test;
 import org.kie.server.api.model.ReleaseId;
 import org.kie.server.api.model.instance.WorkItemInstance;
-import org.kie.server.client.KieServicesException;
+import org.kie.server.api.exception.KieServicesException;
 import org.kie.server.integrationtests.shared.KieServerDeployer;
 
 public class PerProcessInstanceWorkItemIntegrationTest extends JbpmKieServerBaseIntegrationTest {
